requirements.txt
python-dotenv==0.19.2
requests==2.28.1
pandas==1.5.2
